feat: Update app structure and imports
Refactor the main.dart file to import the Home widget from the home_app package instead of the MyCounterApp widget. This change aligns with the new app structure and improves code organization.

Create a new colors.dart file in the custom_widgets folder to define the colors used in the app. This allows for centralized color management and improves code readability.

Add the MainAppBar widget to the custom_widgets folder. This widget represents the main app bar and includes properties for title, elevation, and actions. It implements the PreferredSizeWidget to ensure proper sizing.

Update the README.md file to include a new section for Design and Animations. This section will cover Flutter's design and animation features, providing developers with the tools to create visually stunning and dynamic user interfaces.

Recent User Commits:
- readme updatess
- creation of my counter app
- created the mini projects folder with myCounterApp
- introduction to animations
- UI Components section
- i have learned some new concepts under UIComponents
- REAMME file update
- REAMME file update
- REAMME file update
- REAMME file update

Recent Repository Commits:
- readme updatess
- creation of my counter app
- created the mini projects folder with myCounterApp
- introduction to animations
- UI Components section
- i have learned some new concepts under UIComponents